Extract RaffleListItem
